Add files via upload
This is a website template is inspired by tech influencers. This website has the necessities for any influencers other than tech as well. In the display you can see different tab pages (blog, social, about, contact) in order for the influencer to be able to connect with the audience and share their opinions. There are currently no links linked to the template. Purpose of this template is to display the use of HTML & CSS.